load photo details

+++ app/src/main/java/works/pel/madgallery/photos/api/model/Photo.kt
@@ -1,8 +1,8 @@
 package works.pel.madgallery.photos.api.model
 
 data class Photo(
-    val albumId: Int,
-    val id: Int,
+    val albumId: Long,
+    val id: Long,
     val thumbnailUrl: String,
     val title: String,
     val url: String
